requirements.txt
easydict==1.10
jittor==1.3.8.5
matplotlib==3.5.3
numba==0.56.4
numpy==1.21.6
open3d==0.17.0
opencv_python==4.6.0.66
prefetch_generator==1.0.1
progressbar33==2.4
pyvista==0.37.0
PyYAML==6.0
PyYAML==6.0.1
scipy==1.7.3
tqdm==4.64.1
trimesh==3.23.0
